+++ crates/dapf/src/lib.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 pub mod acceptance;
+pub mod cli_parsers;
 pub mod http_client;
 mod test_durations;
 pub mod test_routes;
