+++ crates/dapf/src/main.rs
@@ -2,19 +2,20 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 use anyhow::{anyhow, bail, Context, Result};
-use clap::{builder::PossibleValue, Args, Parser, Subcommand, ValueEnum};
+use clap::{Args, Parser, Subcommand};
 use dapf::{
     acceptance::{load_testing, LoadControlParams, LoadControlStride, TestOptions},
+    cli_parsers::{
+        use_or_request_from_user, use_or_request_from_user_or_default, CliDapQueryConfig,
+        CliHpkeKemId, CliTaskId, CliVdafConfig,
+    },
     deduce_dap_version_from_url, response_to_anyhow, HttpClient,
 };
 use daphne::{
     constants::DapMediaType,
     error::aborts::ProblemDetails,
-    hpke::{HpkeKemId, HpkeReceiverConfig},
-    messages::{
-        self, encode_base64url, Base64Encode, BatchSelector, Collection, CollectionReq, Query,
-        TaskId,
-    },
+    hpke::HpkeReceiverConfig,
+    messages::{self, encode_base64url, BatchSelector, Collection, CollectionReq, Query, TaskId},
     vdaf::{Prio3Config, VdafConfig},
     DapAggregationParam, DapMeasurement, DapQueryConfig, DapVersion,
 };
@@ -25,192 +26,15 @@ use daphne_service_utils::{
 };
 use prio::codec::{ParameterizedDecode, ParameterizedEncode};
 use rand::{thread_rng, Rng};
-use std::fmt;
 use std::{
-    io::{self, stdin, IsTerminal, Read},
-    ops::ControlFlow,
+    io::{stdin, IsTerminal, Read},
     path::PathBuf,
     process::Command,
-    str::FromStr,
     time::{Duration, SystemTime},
 };
 use tracing::level_filters::LevelFilter;
 use tracing_subscriber::EnvFilter;
 use url::Url;
-
-#[derive(Debug, Clone, Copy)]
-enum DefaultVdafConfigs {
-    Prio2Dimension99k,
-    Prio3NumProofs2,
-    Prio3NumProofs3,
-}
-
-impl fmt::Display for DefaultVdafConfigs {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(self.to_possible_value().unwrap().get_name())
-    }
-}
-
-impl ValueEnum for DefaultVdafConfigs {
-    fn value_variants<'a>() -> &'a [Self] {
-        &[
-            Self::Prio2Dimension99k,
-            Self::Prio3NumProofs2,
-            Self::Prio3NumProofs3,
-        ]
-    }
-
-    fn to_possible_value(&self) -> Option<PossibleValue> {
-        match self {
-            Self::Prio2Dimension99k => Some(PossibleValue::new("prio2-dimension-99k")),
-            Self::Prio3NumProofs2 => Some(PossibleValue::new("prio3-num-proofs-2")),
-            Self::Prio3NumProofs3 => Some(PossibleValue::new("prio3-num-proofs-3")),
-        }
-    }
-}
-
-impl DefaultVdafConfigs {
-    fn into_vdaf(self) -> VdafConfig {
-        match self {
-            Self::Prio2Dimension99k => VdafConfig::Prio2 { dimension: 99_992 },
-            Self::Prio3NumProofs2 => {
-                VdafConfig::Prio3(Prio3Config::SumVecField64MultiproofHmacSha256Aes128 {
-                    bits: 1,
-                    length: 100_000,
-                    chunk_length: 320,
-                    num_proofs: 2,
-                })
-            }
-            Self::Prio3NumProofs3 => {
-                VdafConfig::Prio3(Prio3Config::SumVecField64MultiproofHmacSha256Aes128 {
-                    bits: 1,
-                    length: 100_000,
-                    chunk_length: 320,
-                    num_proofs: 3,
-                })
-            }
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-enum CliVdafConfig {
-    Default(DefaultVdafConfigs),
-    Custom(VdafConfig),
-}
-
-impl fmt::Display for CliVdafConfig {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Default(default) => write!(f, "{default}"),
-            Self::Custom(custom) => write!(f, "{custom}"),
-        }
-    }
-}
-
-impl FromStr for CliVdafConfig {
-    type Err = String;
-    fn from_str(s: &str) -> std::prelude::v1::Result<Self, Self::Err> {
-        DefaultVdafConfigs::from_str(s, false)
-            .map(Self::Default)
-            .or_else(|default_err| {
-                serde_json::from_str(s)
-                    .map(Self::Custom)
-                    .map_err(|json_err| {
-                        if s.contains('{') {
-                            json_err.to_string()
-                        } else {
-                            default_err
-                        }
-                    })
-            })
-    }
-}
-
-impl CliVdafConfig {
-    fn into_vdaf(self) -> VdafConfig {
-        match self {
-            Self::Default(d) => d.into_vdaf(),
-            Self::Custom(v) => v,
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-struct CliDapQueryConfig(DapQueryConfig);
-
-impl fmt::Display for CliDapQueryConfig {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        <DapQueryConfig as fmt::Display>::fmt(&self.0, f)
-    }
-}
-
-impl From<CliDapQueryConfig> for DapQueryConfig {
-    fn from(CliDapQueryConfig(val): CliDapQueryConfig) -> Self {
-        val
-    }
-}
-
-impl From<DapQueryConfig> for CliDapQueryConfig {
-    fn from(value: DapQueryConfig) -> Self {
-        Self(value)
-    }
-}
-
-impl FromStr for CliDapQueryConfig {
-    type Err = String;
-    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
-        if s == "time-interval" {
-            Ok(Self(DapQueryConfig::TimeInterval))
-        } else if let Some(size) = s.strip_prefix("fixed-size") {
-            Ok(Self(DapQueryConfig::FixedSize {
-                max_batch_size: if let Some(size) = size.strip_prefix("-") {
-                    Some(
-                        size.parse()
-                            .map_err(|e| format!("{s} is an invalid query config: {e:?}"))?,
-                    )
-                } else if size.is_empty() {
-                    None
-                } else {
-                    return Err(format!("{size} is an invalid fixed size max batch size"));
-                },
-            }))
-        } else {
-            Err(format!("{s} is an invalid query config"))
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-struct CliTaskId(TaskId);
-
-impl fmt::Display for CliTaskId {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        <TaskId as fmt::Display>::fmt(&self.0, f)
-    }
-}
-
-impl From<TaskId> for CliTaskId {
-    fn from(id: TaskId) -> Self {
-        Self(id)
-    }
-}
-
-impl From<CliTaskId> for TaskId {
-    fn from(CliTaskId(id): CliTaskId) -> Self {
-        id
-    }
-}
-
-impl FromStr for CliTaskId {
-    type Err = anyhow::Error;
-    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
-        TaskId::try_from_base64url(s)
-            .ok_or_else(|| anyhow!("failed to decode ID"))
-            .context("expected URL-safe, base64 string")
-            .map(Self)
-    }
-}
 
 /// DAP Functions, a utility for interacting with DAP deployments.
 #[derive(Parser, Debug)]
@@ -353,7 +177,7 @@ enum HpkeAction {
     /// first by prio and then in base64. This version can later be used when TODO
     Generate {
         #[arg(default_value_t)]
-        kem_alg: KemAlg,
+        kem_alg: CliHpkeKemId,
     },
     /// Rotate the HPKE config advertised by the Aggregator.
     RotateReceiverConfig {
@@ -364,7 +188,7 @@ enum HpkeAction {
         #[arg(default_value_t, long)]
         dap_version: DapVersion,
         #[arg(default_value_t, long)]
-        kem_alg: KemAlg,
+        kem_alg: CliHpkeKemId,
     },
 }
 
@@ -375,7 +199,7 @@ enum TestAction {
     AddHpkeConfig {
         aggregator_url: Url,
         #[arg(short, long, default_value_t)]
-        kem_alg: KemAlg,
+        kem_alg: CliHpkeKemId,
     },
     CreateAddTaskJson {
         #[arg(long)]
@@ -409,7 +233,7 @@ enum TestAction {
         /// If a new hpke config should be inserted after clearing all storage, use this flag to
         /// specify the key algorithm to use for the new hpke config
         #[arg(short, long)]
-        set_hpke_key: Option<KemAlg>,
+        set_hpke_key: Option<CliHpkeKemId>,
     },
 }
 
@@ -513,38 +337,6 @@ impl LoadControlParamsCli {
                 (None, _) => None,
             },
         )
-    }
-}
-
-#[derive(Clone, Debug)]
-struct KemAlg(HpkeKemId);
-
-impl Default for KemAlg {
-    fn default() -> Self {
-        Self(HpkeKemId::X25519HkdfSha256)
-    }
-}
-
-impl fmt::Display for KemAlg {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.to_possible_value().unwrap().get_name())
-    }
-}
-
-impl ValueEnum for KemAlg {
-    fn value_variants<'a>() -> &'a [Self] {
-        &[
-            Self(HpkeKemId::X25519HkdfSha256),
-            Self(HpkeKemId::P256HkdfSha256),
-        ]
-    }
-
-    fn to_possible_value(&self) -> Option<PossibleValue> {
-        Some(match self.0 {
-            HpkeKemId::X25519HkdfSha256 => PossibleValue::new("x25519_hkdf_sha256"),
-            HpkeKemId::P256HkdfSha256 => PossibleValue::new("p256_hkdf_sha256"),
-            HpkeKemId::NotImplemented(id) => unreachable!("unhandled HPKE KEM ID {id}"),
-        })
     }
 }
 
@@ -1101,54 +893,52 @@ async fn handle_test_routes(action: TestAction, http_client: HttpClient) -> anyh
             time_precision,
             expires_in_seconds: task_expiration,
         } => {
-            let (vdaf_verify_key, vdaf) = use_or_request_from_user_or_default(
-                vdaf,
-                || CliVdafConfig::Default(DefaultVdafConfigs::Prio3NumProofs3),
-                "vdaf",
-            )
-            .map(|vdaf| {
-                let vdaf = vdaf.into_vdaf();
-                let vdaf_verify_key = vdaf.gen_verify_key();
-                let (typ, bits, length, chunk_length) = match vdaf {
-                    VdafConfig::Prio3(prio3) => match prio3 {
-                        Prio3Config::Count => ("Prio3Count", None, None, None),
-                        Prio3Config::Sum { bits } => ("Prio3Sum", Some(bits), None, None),
-                        Prio3Config::Histogram {
-                            length,
-                            chunk_length,
-                        } => ("Prio3Histogram", None, Some(length), Some(chunk_length)),
-                        Prio3Config::SumVec {
-                            bits,
-                            length,
-                            chunk_length,
-                        } => ("Prio3SumVec", Some(bits), Some(length), Some(chunk_length)),
-                        Prio3Config::SumVecField64MultiproofHmacSha256Aes128 {
-                            bits,
-                            length,
-                            chunk_length,
-                            num_proofs: _,
-                        } => (
-                            "Prio3SumVecField64MultiproofHmacSha256Aes128",
-                            Some(bits),
-                            Some(length),
-                            Some(chunk_length),
-                        ),
+            let (vdaf_verify_key, vdaf) =
+                use_or_request_from_user_or_default(vdaf, CliVdafConfig::default, "vdaf").map(
+                    |vdaf| {
+                        let vdaf = vdaf.into_vdaf();
+                        let vdaf_verify_key = vdaf.gen_verify_key();
+                        let (typ, bits, length, chunk_length) = match vdaf {
+                            VdafConfig::Prio3(prio3) => match prio3 {
+                                Prio3Config::Count => ("Prio3Count", None, None, None),
+                                Prio3Config::Sum { bits } => ("Prio3Sum", Some(bits), None, None),
+                                Prio3Config::Histogram {
+                                    length,
+                                    chunk_length,
+                                } => ("Prio3Histogram", None, Some(length), Some(chunk_length)),
+                                Prio3Config::SumVec {
+                                    bits,
+                                    length,
+                                    chunk_length,
+                                } => ("Prio3SumVec", Some(bits), Some(length), Some(chunk_length)),
+                                Prio3Config::SumVecField64MultiproofHmacSha256Aes128 {
+                                    bits,
+                                    length,
+                                    chunk_length,
+                                    num_proofs: _,
+                                } => (
+                                    "Prio3SumVecField64MultiproofHmacSha256Aes128",
+                                    Some(bits),
+                                    Some(length),
+                                    Some(chunk_length),
+                                ),
+                            },
+                            VdafConfig::Prio2 { .. } => ("Prio2", None, None, None),
+                            VdafConfig::Pine(_) => ("Pine", None, None, None),
+                            #[cfg(feature = "experimental")]
+                            VdafConfig::Mastic { .. } => todo!(),
+                        };
+                        (
+                            encode_base64url(vdaf_verify_key),
+                            InternalTestVdaf {
+                                typ: typ.into(),
+                                bits: bits.map(|a| a.to_string()),
+                                length: length.map(|a| a.to_string()),
+                                chunk_length: chunk_length.map(|a| a.to_string()),
+                            },
+                        )
                     },
-                    VdafConfig::Prio2 { .. } => ("Prio2", None, None, None),
-                    VdafConfig::Pine(_) => ("Pine", None, None, None),
-                    #[cfg(feature = "experimental")]
-                    VdafConfig::Mastic { .. } => todo!(),
-                };
-                (
-                    encode_base64url(vdaf_verify_key),
-                    InternalTestVdaf {
-                        typ: typ.into(),
-                        bits: bits.map(|a| a.to_string()),
-                        length: length.map(|a| a.to_string()),
-                        chunk_length: chunk_length.map(|a| a.to_string()),
-                    },
-                )
-            })?;
+                )?;
             let CliDapQueryConfig(query) = use_or_request_from_user_or_default(
                 query,
                 || DapQueryConfig::FixedSize {
@@ -1223,98 +1013,6 @@ async fn handle_test_routes(action: TestAction, http_client: HttpClient) -> anyh
             };
 
             Ok(())
-        }
-    }
-}
-
-fn use_or_request_from_user_or_default<T, U>(
-    value: Option<T>,
-    mut default: impl FnMut() -> U,
-    field: &'static str,
-) -> io::Result<T>
-where
-    U: Into<T>,
-    T: FromStr + fmt::Display,
-    T::Err: fmt::Debug,
-{
-    match value {
-        Some(v) => Ok(v),
-        None => Ok(request_from_user::<T>(
-            field,
-            format!("leave empty to default: {}", default().into()).as_str(),
-            || ControlFlow::Break(Some(default().into())),
-        )?
-        .unwrap()),
-    }
-}
-
-fn use_or_request_from_user<T>(value: Option<T>, field: &'static str) -> io::Result<T>
-where
-    T: FromStr + fmt::Display,
-    T::Err: fmt::Debug,
-{
-    match value {
-        Some(v) => Ok(v),
-        None => request_from_user::<T>(field, None, || ControlFlow::Continue(()))
-            .transpose()
-            .unwrap()
-            .map(Into::into),
-    }
-}
-
-fn request_from_user<'e, T>(
-    field: &'static str,
-    extra_info: impl Into<Option<&'e str>>,
-    mut on_empty: impl FnMut() -> ControlFlow<Option<T>>,
-) -> io::Result<Option<T>>
-where
-    T: FromStr + fmt::Display,
-    T::Err: fmt::Debug,
-{
-    if !std::io::stdin().is_terminal() {
-        return match on_empty() {
-            ControlFlow::Break(b) => {
-                eprintln!(
-                    "defaulting {field} to {}",
-                    b.as_ref()
-                        .map(|b| format!("{b}"))
-                        .unwrap_or_else(|| "None".into())
-                );
-                Ok(b)
-            }
-            ControlFlow::Continue(()) => Err(io::Error::other(format!(
-                "{field:?} is required. use command line options to specify it"
-            ))),
-        };
-    }
-    let extra_info = extra_info.into();
-    loop {
-        eprint!(
-            "{field}{}?: ",
-            extra_info.map(|e| format!(" ({e})")).unwrap_or_default()
-        );
-        let mut buf = String::new();
-        std::io::stdin().read_line(&mut buf)?;
-
-        if buf.is_empty() {
-            // <C-d>
-            break Err(io::ErrorKind::UnexpectedEof.into());
-        }
-
-        buf.pop();
-
-        if buf.is_empty() {
-            match on_empty() {
-                ControlFlow::Continue(()) => continue,
-                ControlFlow::Break(b) => break Ok(b),
-            }
-        }
-
-        match buf.parse::<T>() {
-            Ok(t) => break Ok(Some(t)),
-            Err(e) => {
-                eprintln!("{buf} is not a valid {field}: {e:?}");
-            }
         }
     }
 }
